+++ src/language/scala.rs
@@ -140,6 +140,7 @@ static DEFINITION: LanguageDef = LanguageDef {
     common_types: COMMON_TYPES,
     container_body_kinds: &["template_body"],
     extract_container_name: None,
+    extract_qualified_method: None,
 };
 
 pub fn definition() -> &'static LanguageDef {